requirements.txt
streamlit>=1.35.0
supabase>=2.4.0
qrcode[pil]>=7.4.2
Pillow>=10.0.0
fpdf2>=2.7.9
python-dotenv>=1.0.0
